moved isfinite method to the Array API submodule, with the correct formatting.
