+++ ivy/functional/backends/mxnet/array_api/elementwise_functions.py
@@ -1,0 +1,11 @@
+# global
+import mxnet as mx
+
+# local
+from ivy.functional.backends.mxnet import _handle_flat_arrays_in_out
+
+
+@_handle_flat_arrays_in_out
+def isfinite(x: mx.ndarray.ndarray.NDArray) -> mx.ndarray.ndarray.NDArray:
+    # ToDo: remove float32 conversion once int8 and uint8 work correctly. Currently 0 returns 0 for these types.
+    return mx.nd.contrib.isfinite(x.astype('float32')).astype('bool')
