+++ ivy/functional/backends/mxnet/core/general.py
@@ -61,61 +61,6 @@ DTYPE_FROM_STR = {'int8': _np.int8,
                 'bool': _np.bool_}
 
 
-# Helpers #
-# --------#
-
-def _raise(ex):
-    raise ex
-
-
-def _mxnet_init_context(dev):
-    dev = dev_to_str(dev)
-    if dev is None or dev.find("cpu") != -1:
-        mx_dev = "cpu"
-    elif dev.find("gpu") != -1:
-        mx_dev = "gpu"
-    else:
-        raise Exception("dev input {} not supported.".format(dev))
-    if dev.find(":") != -1:
-        mx_dev_id = int(dev[dev.find(":")+1:])
-    else:
-        mx_dev_id = 0
-    return _mx.Context(mx_dev, mx_dev_id)
-
-
-def _scalar_or_flat_array_to_scalar(x):
-    return x if isinstance(x, Number) else (x.asscalar() if len(x.shape) == 0 else x)
-
-
-def _flat_array_to_1_dim_array(x):
-    return _mx.nd.array([x.asscalar()]).astype(dtype(x)) if len(x.shape) == 0 else x
-
-
-def _1_dim_array_to_flat_array(x):
-    return _mx.nd.array(x.asscalar(), dtype=x.dtype) if x.shape == (1,) else x
-
-
-def _handle_flat_arrays_in(fn):
-    return _handle_flat_arrays_in_out(fn, False)
-
-
-def _handle_flat_arrays_in_out(fn, include_out=True):
-    def wrapped_fn(*args, **kwargs):
-        expanded = False
-        def expand(x):
-            nonlocal expanded
-            expanded = True
-            return _flat_array_to_1_dim_array(x)
-
-        args_expanded = ivy.nested_map(args, lambda x: expand(x) if ivy.is_array(x) and len(x.shape) == 0 else x)
-        kwargs_expanded = ivy.nested_map(kwargs, lambda x: expand(x) if ivy.is_array(x) and len(x.shape) == 0 else x)
-        ret = fn(*args_expanded, **kwargs_expanded)
-        if expanded and include_out:
-            return ivy.nested_map(ret, lambda x: _1_dim_array_to_flat_array(x) if ivy.is_array(x) else x)
-        return ret
-    return wrapped_fn
-
-
 # API #
 # ----#
 
@@ -413,12 +358,6 @@ def isnan(x):
 @_handle_flat_arrays_in_out
 def isinf(x):
     return _mx.nd.contrib.isinf(x).astype('bool')
-
-
-@_handle_flat_arrays_in_out
-def isfinite(x):
-    # ToDo: remove float32 conversion once int8 and uint8 work correctly. Currently 0 returns 0 for these types.
-    return _mx.nd.contrib.isfinite(x.astype('float32')).astype('bool')
 
 
 reshape = lambda x, new_shape: x.reshape(new_shape)
