+++ ivy/functional/backends/numpy/array_api/elementwise_functions.py
@@ -1,0 +1,6 @@
+# global
+import numpy as np
+
+
+def isfinite(x: np.ndarray) -> np.ndarray:
+    return np.isfinite(x)
