+++ ivy/functional/backends/torch/array_api/elementwise_functions.py
@@ -1,0 +1,5 @@
+# global
+import torch
+
+def isfinite(x: torch.Tensor) -> torch.Tensor:
+    return torch.isfinite(x)
