+++ ivy/functional/backends/torch/core/general.py
@@ -381,10 +381,6 @@ def isinf(x):
     return _torch.isinf(x)
 
 
-def isfinite(x):
-    return _torch.isfinite(x)
-
-
 def reshape(x, newshape: List[int]):
     if isinstance(newshape, int):
         newshape = [newshape]
